Add files via upload
This notebook focuses on filtering the augmented dataset of disease-related FAQs by removing all entries that contain the original dataset's contents. This ensures that the filtered augmented dataset is used for training, while the unseen original dataset is reserved for evaluation, thereby improving the model's ability to generalize to new data.