Orginal Dataset
This file contains the original dataset of disease-related FAQs, comprising question-answer pairs about various health conditions. The dataset was gathered via web scraping and serves as the primary evaluation dataset in the project. It includes a wide range of health topics, designed to assist in fine-tuning language models like Llama 3.1 for better question-answering capabilities in the medical domain. This dataset remains untouched during the training process, making it ideal for evaluating model performance on unseen data.